refactor: view -> InputView, OutputView로 분리
